requirements.txt
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
Pillow-SIMD==7.0.0.post3
tqdm==4.56.0
